#485 rename card_resolver to card-resolver and ensure Vary header is preserved on agent card 500 errors
