+++ src/compat/v0_3/client/index.ts
@@ -2,7 +2,7 @@
  * Compat-layer client components for v0.3 transports.
  */
 
-export { isLegacyAgentCard, parseLegacyAgentCard } from './card_resolver.js';
+export { isLegacyAgentCard, parseLegacyAgentCard } from './card-resolver.js';
 export {
   LegacyJsonRpcTransport,
   type LegacyJsonRpcTransportOptions,
